feat: add full-page-modal component
This modal will be used to show the contents of different app-card's
"screen" content.

It was unclear from the instructions of the task what "screen" means. From
the absence of a changed url in mockups, and the mention of "cancel button
should show previous screen" in description of task, I assumed that the "screen"
means some sort of modal. So it is a modal now.